Refactor fakes param parsing
There was too much reliance on convention which duplicated logic,
incorrectly in some cases. Now, any params that required parsing are
returned from parseHeaderPathQueryParams.
The "final" parsed param local vars now all have the suffix Param.
Removed some superfluous casts in the fake invocation call.
Removed some unnecessary local vars for values that are strings.
Fixed bad codegen for enums that aren't string values.
Added more test coverage for various scenarios.

+++ packages/autorest.go/test/autorest/optionalgroup/fake/zz_explicit_server.go
@@ -649,7 +649,7 @@ func (e *ExplicitServerTransport) dispatchPostRequiredIntegerHeader(req *http.Re
 	if err != nil {
 		return nil, err
 	}
-	respr, errRespr := e.srv.PostRequiredIntegerHeader(req.Context(), int32(headerParameterParam), nil)
+	respr, errRespr := e.srv.PostRequiredIntegerHeader(req.Context(), headerParameterParam, nil)
 	if respErr := server.GetError(errRespr, req); respErr != nil {
 		return nil, respErr
 	}
